Initial commit of the uniform generator
I wanted to get something in the repo so that Tom could see what it is that
we are going after with this approach. The idea being to make sequence
generators that can then be used to create other generators, etc. It's all
about the composition and lazy sequences in clojure.