Added the mapper generator to enable wild stuff
We really needed to have a way to map the reandom variables into something
else, and so I whipped this up and then tested it. It allows us to make any
kind of mapping we need and from this, new random variables - based on the
underlying distribution, and the mapping data.