Updated to latest Gorilla REPL
I wanted to have a look at the latest Gorilla REPL, and it's changed a
lot, which is really nice to know. The main work is on another
maintainer, but it loads all the old workspaces, and with just a few
changes to the versions of the dependencies, it's all working just fine.